requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=hd77b12b_0
absl-py=1.4.0=pypi_0
altair=4.1.0=py_1
arrow-cpp=11.0.0=py39h3577439_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=py39haa95532_0
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39haa95532_0
boost-cpp=1.73.0=h2bbff1b_12
boto3=1.26.164=pypi_0
botocore=1.29.164=pypi_0
bottleneck=1.3.5=py39h080aedc_0
brotlipy=0.7.0=py39h2bbff1b_1003
bzip2=1.0.8=he774522_0
c-ares=1.19.0=h2bbff1b_0
ca-certificates=2023.05.30=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.5.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39haa95532_0
colorama=0.4.6=py39haa95532_0
comm=0.1.2=py39haa95532_0
contourpy=1.1.0=pypi_0
cryptography=39.0.1=py39h21b164f_0
cycler=0.10.0=pypi_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.12.2=pypi_0
fire=0.5.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.40.0=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2023.6.0=pypi_0
gast=0.4.0=pypi_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py39haa95532_0
glog=0.5.0=hd77b12b_0
google-auth=2.21.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.56.0=pypi_0
h5py=3.9.0=pypi_0
huggingface-hub=0.15.1=pypi_0
idna=2.10=pypi_0
importlib-metadata=6.0.0=py39haa95532_0
importlib-resources=5.12.0=pypi_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.19.2=py39hd4e2768_0
ipython=8.12.0=py39haa95532_0
ipywidgets=8.0.4=py39haa95532_0
jax=0.4.13=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.1.2=py39haa95532_0
jmespath=1.0.1=pypi_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py39haa95532_0
jupyter_client=8.1.0=py39haa95532_0
jupyter_core=5.3.0=py39haa95532_0
jupyterlab_widgets=3.0.5=py39haa95532_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_12
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=16.0.0=pypi_0
libcurl=8.1.1=h86230a5_1
libdeflate=1.17=h2bbff1b_0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=hcd4344a_2
libthrift=0.15.0=he1d8c1a_0
libtiff=4.5.0=h6c2663c_2
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=py39haa95532_1
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=py39haa95532_0
mdurl=0.1.0=py39haa95532_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.6=py39hf11a4ad_1
mkl_random=1.2.2=py39hf11a4ad_1
ml-dtypes=0.2.0=pypi_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.5.6=py39haa95532_0
networkx=3.1=pypi_0
numexpr=2.8.4=py39h7b80656_1
numpy=1.25.0=py39h055cbcc_0
numpy-base=1.25.0=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.74=pypi_0
openssl=1.1.1u=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
orc=1.7.4=h623e30f_1
packaging=23.0=py39haa95532_0
pandas=1.5.3=py39hf11a4ad_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39hd77b12b_0
pip=23.1.2=py39haa95532_0
platformdirs=2.5.2=py39haa95532_0
prompt-toolkit=3.0.36=py39haa95532_0
protobuf=3.20.3=py39hd77b12b_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=py39haa45b5f_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybboxes=0.1.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py39haa95532_0
pygments=2.15.1=py39haa95532_1
pympler=0.9=py_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39haa95532_0
python=3.9.13=h6244533_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
pytz=2022.7=py39haa95532_0
pywin32=305=py39h2bbff1b_0
pyyaml=6.0=py39h2bbff1b_1
pyzmq=25.1.0=py39hd77b12b_0
re2=2022.04.01=hd77b12b_0
requests=2.29.0=py39haa95532_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.3.5=py39haa95532_0
roboflow=1.1.0=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.1=pypi_0
sahi=0.11.14=pypi_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
semver=2.13.0=pyhd3eb1b0_0
setuptools=67.8.0=py39haa95532_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.24.0=pypi_0
supervision=0.11.1=pypi_0
sympy=1.12=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=pypi_0
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
terminaltables=3.1.10=pypi_0
tflearn=0.5.0=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py39haa95532_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.2=py39h2bbff1b_0
traitlets=5.7.1=py39haa95532_0
typing_extensions=4.6.3=py39haa95532_0
tzdata=2023c=h04d1e81_0
tzlocal=2.1=py39haa95532_1
ultralytics=8.0.124=pypi_0
urllib3=1.26.16=py39haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py39haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.3.6=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py39haa95532_0
widgetsnbextension=4.0.5=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
word2number=1.1=pypi_0
wrapt=1.14.1=pypi_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
yolov5=7.0.12=pypi_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
